backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiomysql>=0.3.2",
    "fastapi>=0.135.3",
    "passlib[bcrypt]>=1.7.4",
    "PyJWT>=2.10.1",
    "python-dotenv>=1.2.2",
    "sqlalchemy>=2.0.49",
    "supabase>=2.28.3",
    "uvicorn[standard]>=0.44.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
]
[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
